requirements.txt
streamlit
pandas
python-dotenv
openai>=1.12.0